- puts get project in common between upgrade and exteranlize
