+++ src/o365/spfx/commands/project/project-upgrade.ts
@@ -5,13 +5,14 @@ import Command, {
 import GlobalOptions from '../../../../GlobalOptions';
 import * as path from 'path';
 import * as fs from 'fs';
-import { Finding, Utils, Hash, Dictionary } from './project-upgrade/';
+import { Finding, Hash, Dictionary } from './project-upgrade/';
 import { Rule } from './project-upgrade/rules/Rule';
 import * as os from 'os';
-import { Project, Manifest, TsFile, ScssFile } from './project-upgrade/model';
+import { Project } from './project-upgrade/model';
 import { FindingToReport } from './project-upgrade/FindingToReport';
 import { FN017001_MISC_npm_dedupe } from './project-upgrade/rules/FN017001_MISC_npm_dedupe';
 import { ReportData, ReportDataModification } from './ReportData';
+import { BaseProjectCommand } from './base-project-command';
 
 const vorpal: Vorpal = require('../../../../vorpal-init');
 
@@ -26,7 +27,7 @@ interface Options extends GlobalOptions {
   shell?: string;
 }
 
-class SpfxProjectUpgradeCommand extends Command {
+class SpfxProjectUpgradeCommand extends BaseProjectCommand {
   private projectVersion: string | undefined;
   private toVersion: string = '';
   private packageManager: string = 'npm';
@@ -361,162 +362,6 @@ class SpfxProjectUpgradeCommand extends Command {
     else {
       cmd.log(findingsToReport);
     }
-  }
-
-  private getProject(projectRootPath: string): Project {
-    const project: Project = {
-      path: projectRootPath
-    };
-
-    const configJsonPath: string = path.join(projectRootPath, 'config/config.json');
-    if (fs.existsSync(configJsonPath)) {
-      try {
-        project.configJson = JSON.parse(Utils.removeSingleLineComments(fs.readFileSync(configJsonPath, 'utf-8')));
-      }
-      catch { }
-    }
-
-    const copyAssetsJsonPath: string = path.join(projectRootPath, 'config/copy-assets.json');
-    if (fs.existsSync(copyAssetsJsonPath)) {
-      try {
-        project.copyAssetsJson = JSON.parse(Utils.removeSingleLineComments(fs.readFileSync(copyAssetsJsonPath, 'utf-8')));
-      }
-      catch { }
-    }
-
-    const deployAzureStorageJsonPath: string = path.join(projectRootPath, 'config/deploy-azure-storage.json');
-    if (fs.existsSync(deployAzureStorageJsonPath)) {
-      try {
-        project.deployAzureStorageJson = JSON.parse(Utils.removeSingleLineComments(fs.readFileSync(deployAzureStorageJsonPath, 'utf-8')));
-      }
-      catch { }
-    }
-
-    const packageJsonPath: string = path.join(projectRootPath, 'package.json');
-    if (fs.existsSync(packageJsonPath)) {
-      try {
-        project.packageJson = JSON.parse(Utils.removeSingleLineComments(fs.readFileSync(packageJsonPath, 'utf-8')));
-      }
-      catch { }
-    }
-
-    const packageSolutionJsonPath: string = path.join(projectRootPath, 'config/package-solution.json');
-    if (fs.existsSync(packageSolutionJsonPath)) {
-      try {
-        project.packageSolutionJson = JSON.parse(Utils.removeSingleLineComments(fs.readFileSync(packageSolutionJsonPath, 'utf-8')));
-      }
-      catch { }
-    }
-
-    const serveJsonPath: string = path.join(projectRootPath, 'config/serve.json');
-    if (fs.existsSync(serveJsonPath)) {
-      try {
-        project.serveJson = JSON.parse(Utils.removeSingleLineComments(fs.readFileSync(serveJsonPath, 'utf-8')));
-      }
-      catch { }
-    }
-
-    const tsConfigJsonPath: string = path.join(projectRootPath, 'tsconfig.json');
-    if (fs.existsSync(tsConfigJsonPath)) {
-      try {
-        project.tsConfigJson = JSON.parse(Utils.removeSingleLineComments(fs.readFileSync(tsConfigJsonPath, 'utf-8')));
-      }
-      catch { }
-    }
-
-    const tsLintJsonPath: string = path.join(projectRootPath, 'config/tslint.json');
-    if (fs.existsSync(tsLintJsonPath)) {
-      try {
-        project.tsLintJson = JSON.parse(Utils.removeSingleLineComments(fs.readFileSync(tsLintJsonPath, 'utf-8')));
-      }
-      catch { }
-    }
-
-    const tsLintJsonRootPath: string = path.join(projectRootPath, 'tslint.json');
-    if (fs.existsSync(tsLintJsonRootPath)) {
-      try {
-        project.tsLintJsonRoot = JSON.parse(Utils.removeSingleLineComments(fs.readFileSync(tsLintJsonRootPath, 'utf-8')));
-      }
-      catch { }
-    }
-
-    const writeManifestJsonPath: string = path.join(projectRootPath, 'config/write-manifests.json');
-    if (fs.existsSync(writeManifestJsonPath)) {
-      try {
-        project.writeManifestsJson = JSON.parse(Utils.removeSingleLineComments(fs.readFileSync(writeManifestJsonPath, 'utf-8')));
-      }
-      catch { }
-    }
-
-    const yoRcJsonPath: string = path.join(projectRootPath, '.yo-rc.json');
-    if (fs.existsSync(yoRcJsonPath)) {
-      try {
-        project.yoRcJson = JSON.parse(Utils.removeSingleLineComments(fs.readFileSync(yoRcJsonPath, 'utf-8')));
-      }
-      catch { }
-    }
-
-    const manifests: Manifest[] = [];
-    const files: string[] = Utils.getAllFiles(path.join(projectRootPath, 'src')) as string[];
-    files.forEach(f => {
-      if (f.endsWith('.manifest.json')) {
-        try {
-          const manifestStr = Utils.removeSingleLineComments(fs.readFileSync(f, 'utf-8'));
-          const manifest: Manifest = JSON.parse(manifestStr);
-          manifest.path = f;
-          manifests.push(manifest);
-        }
-        catch { }
-      }
-    });
-    project.manifests = manifests;
-
-    const gulpfileJsPath: string = path.join(projectRootPath, 'gulpfile.js');
-    if (fs.existsSync(gulpfileJsPath)) {
-      project.gulpfileJs = {
-        src: fs.readFileSync(gulpfileJsPath, 'utf-8')
-      };
-    }
-
-    project.vsCode = {};
-    const vsCodeSettingsPath: string = path.join(projectRootPath, '.vscode', 'settings.json');
-    if (fs.existsSync(vsCodeSettingsPath)) {
-      try {
-        project.vsCode.settingsJson = JSON.parse(Utils.removeSingleLineComments(fs.readFileSync(vsCodeSettingsPath, 'utf-8')));
-      }
-      catch { }
-    }
-
-    const vsCodeExtensionsPath: string = path.join(projectRootPath, '.vscode', 'extensions.json');
-    if (fs.existsSync(vsCodeExtensionsPath)) {
-      try {
-        project.vsCode.extensionsJson = JSON.parse(Utils.removeSingleLineComments(fs.readFileSync(vsCodeExtensionsPath, 'utf-8')));
-      }
-      catch { }
-    }
-
-    const vsCodeLaunchPath: string = path.join(projectRootPath, '.vscode', 'launch.json');
-    if (fs.existsSync(vsCodeLaunchPath)) {
-      try {
-        project.vsCode.launchJson = JSON.parse(Utils.removeSingleLineComments(fs.readFileSync(vsCodeLaunchPath, 'utf-8')));
-      }
-      catch { }
-    }
-
-    const srcFiles: string[] = SpfxProjectUpgradeCommand.readdirR(path.join(projectRootPath, 'src')) as string[];
-    const tsFiles: string[] = srcFiles.filter(f => f.endsWith('.ts') || f.endsWith('.tsx'));
-    project.tsFiles = tsFiles.map(f => new TsFile(f));
-
-    const scssFiles: string[] = srcFiles.filter(f => f.endsWith('.scss'));
-    project.scssFiles = scssFiles.map(f => new ScssFile(f));
-
-    return project;
-  }
-
-  private static readdirR(dir: string): string | string[] {
-    return fs.statSync(dir).isDirectory()
-      ? Array.prototype.concat(...fs.readdirSync(dir).map(f => SpfxProjectUpgradeCommand.readdirR(path.join(dir, f))))
-      : dir;
   }
 
   private getTextReport(findings: FindingToReport[]): string {
